sem2/app 2/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy>=2.0.25
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.12.1
redis==5.0.1
python-dotenv==1.0.0
websockets==12.0
pydantic-settings
celery==5.3.6 